c3c8e82 | Prevent race condition at Route Mgr start/stop events (rebasing_on_daily_reference) [François Gaffie]
4bb58e8 | Remove some redundant logs [François Gaffie]
8e26016 | Enforce consitancy of setter API within RouteCollection [François Gaffie]
de9badd | Refactor Route Manager by introducing collections [Sebastien Gonzalve]
570a3ec | move mask getters inside struct [Sebastien Gonzalve]
7e3b30d | allocate element outside addElement [Sebastien Gonzalve]
16c1760 | Remove unneeded asserts [François Gaffie]
c0fb7c4 | Remove map from Element accessor APIs [Sebastien Gonzalve]
e1348a6 | Fix Klocwork issues within AudioHAL [François Gaffie]
bc54973 | CriterionParameter takes a reference on criterion [François Gaffie]
ec190ff | Refactor Audio HAL [François Gaffie]
88a2572 | Handle BADF error within OutputStreams. [François Gaffie]
9cd67fa | Revert "[Audio] Handle Bad File Number (EBADF) error in write stream function." [François Gaffie]

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-17002

+++ audio_route_manager/includes/StreamRouteConfig.hpp
@@ -34,7 +34,7 @@ struct StreamRouteConfig
     bool requirePostDisable;
 
     const char *cardName;   /**< Alsa card used by the stream route. */
-    uint8_t deviceId;       /**< Alsa card used by the stream route. */
+    uint32_t deviceId;       /**< Alsa card used by the stream route. */
 
     /**
      * pcm configuration supported by the stream route.
@@ -49,7 +49,7 @@ struct StreamRouteConfig
     uint32_t silenceThreshold;
     uint32_t availMin;
 
-    uint8_t silencePrologInMs; /**< if needed, silence to be appended before valid samples. */
+    uint32_t silencePrologInMs; /**< if needed, silence to be appended before valid samples. */
     uint32_t flagMask; /**< flags supported by this route. To be checked with stream flags. */
     uint32_t useCaseMask; /**< use cases supported by this route. To be checked with stream. */
